refactor: extract clerk org resolution into shared service function
Move Clerk org resolution logic from route handler into a shared service
function, keeping the route handler thin and eliminating duplication with
discoverAndCreateScope.
